refactor(habitat): move boundary taxonomy inputs out of service model
